requirements.txt
Django
ipython
Pillow
psycopg2
tzdata
djangorestframework
pytz
python-dotenv
django-filter
coverage
djangorestframework-simplejwt
drf_yasg
django-cors-headers
requests
stripe
celery
eventlet
redis
django_celery_beat
telegram
flake8